Re-organize model types
