+++ src/api/buildListingEndpoint/models.ts
@@ -1,9 +1,4 @@
-import {
-  QueryParameter,
-  RegexSearchParameter,
-  ListsSearchParameter,
-  SortQueryParameterValue,
-} from '../../queryParameters/models';
+import { QueryParameter } from '../../queryParameters/models';
 
 export interface BuildListingEndpointParameters {
   baseEndpoint?: string;
@@ -28,3 +23,46 @@ export interface SearchParameter {
   regex?: RegexSearchParameter;
   lists?: Array<ListsSearchParameter>;
 }
+
+export interface ListsSearchQueryParameterValue {
+  $and: Array<{ [field: string]: { [field: string]: { $in: Array<string> } } }>;
+}
+
+export interface SortQueryParameterValue {
+  [sortf: string]: string;
+}
+
+export interface RegexSearchParameter {
+  value: string;
+  fields: Array<string>;
+}
+
+export interface ListsSearchParameter {
+  field: string;
+  values: Array<string>;
+}
+
+type SearchPatterns = Array<{ [field: string]: { $rg: string } }>;
+
+export interface OrSearchQueryParameterValue {
+  $or: SearchPatterns;
+}
+
+export interface AndSearchQueryParameterValue {
+  $and: SearchPatterns;
+}
+
+export type RegexSearchQueryParameterValue =
+  | OrSearchQueryParameterValue
+  | AndSearchQueryParameterValue
+  | undefined;
+
+export type SearchQueryParameterValue =
+  | {
+      $and: Array<
+        RegexSearchQueryParameterValue | ListsSearchQueryParameterValue
+      >;
+    }
+  | RegexSearchQueryParameterValue
+  | ListsSearchQueryParameterValue
+  | undefined;
